Replace logic of main.js in the PageController
